requirements.txt
Flask
flask-pymongo
pymongo
ultralytics
torch
torchvision
opencv-python
numpy
python-dotenv
pillow
twilio
flask-login
flask-wtf
email-validator
matplotlib
seaborn
pandas
werkzeug
bcrypt